[runtime] Cleanup and separate runtime code from core.
cudaq/lib/Target has two distinct libraries being built from the sources
there. One is load-bearing for one of the tools that make up the C++
compiler. The other library is only used by the runtime.

This PR removes the second library and moves those source files to the
runtime, folding them into an existing runtime library.
