+++ runtime/cudaq/CMakeLists.txt
@@ -28,6 +28,7 @@ add_library(${LIBRARY_NAME}
     platform/qpu_utils.cpp
     platform/qpu_state.cpp
     platform/quantum_platform.cpp
+    platform/RuntimeEndpoint.cpp
     platform/platform_iface.cpp
     qis/execution_manager_c_api.cpp
     qis/execution_manager.cpp
