+++ runtime/cudaq/platform/RuntimeEndpoint.h
