+++ runtime/cudaq/platform/orca/OrcaRemoteRESTQPU.h
@@ -10,7 +10,7 @@
 
 #include "OrcaExecutor.h"
 #include "common/CompiledModule.h"
-#include "cudaq/Target/CompileTarget.h"
+#include "common/CompileTarget.h"
 #include "cudaq/platform/qpu.h"
 #include "cudaq/utils/cudaq_utils.h"
 #include "cudaq/utils/owning_ptr.h"
