+++ runtime/cudaq/platform/quantum_platform.h
@@ -18,8 +18,8 @@
 #include "common/SampleResult.h"
 #include "common/ThunkInterface.h"
 #include "nvqpp_interface.h"
-#include "cudaq/Target/CompileTarget.h"
-#include "cudaq/Target/RuntimeEndpoint.h"
+#include "common/CompileTarget.h"
+#include "cudaq/platform/RuntimeEndpoint.h"
 #include "cudaq/algorithms/dem/policy.h"
 #include "cudaq/platform/qpu.h"
 #include "cudaq/remote_capabilities.h"
