+++ runtime/test/Regress/compile_target_conditional_feedback_crash.cpp
@@ -16,7 +16,7 @@
 #include "common/KernelArgs.h"
 #include "cudaq_internal/compiler/Compiler.h"
 #include "cudaq_internal/compiler/RuntimeMLIR.h"
-#include "cudaq/Target/CompileTarget.h"
+#include "common/CompileTarget.h"
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/Parser/Parser.h"
