+++ targettests/analog/analog_launch_payload.cpp
@@ -16,7 +16,7 @@
 // RUN: nvq++ %s -o %t && %t | FileCheck %s
 // clang-format on
 
-#include "cudaq/Target/RuntimeEndpoint.h"
+#include "cudaq/platform/RuntimeEndpoint.h"
 #include "cudaq/algorithms/evolve_internal.h"
 #include "cudaq/platform/quantum_platform.h"
 #include <cstdio>
